requirements.txt
Django>=4.2,<5.0
djongo>=1.3,<2.0
djangorestframework>=3.15,<4.0
pymongo>=3.12,<4.0
